+++ src/API/Dashboard/Shows/Slug/Edit/Get/Templates/Form.hs
@@ -80,7 +80,7 @@ template backend showModel userMeta isStaff schedulesJson eligibleHosts currentH
         textareaField "description" 6 $ do
           label "Description"
           placeholder "Describe your show. What kind of music do you play? What's your show's vibe?"
-          value showModel.description
+          maybe (pure ()) value showModel.description
           maxLength 5000
 
         textField "tags" $ do